Add Javascript Function passby value or reference example
Functions such as control flow statements (if, for, while, etc.) can be
initialized by passing arguments into the function body. In JavaScript,
variables are either a complex type (e.g., Object, Array) or a primitive
type (e.g., String, Integer). When a complex object is supplied as an
argument, it is passed by reference to the function body. Instead of
sending a copy of the variable, JavaScript sends a pointer to its
location in the memory heap. Conversely, when passing a primitive type
to a function, JavaScript passes by value. This difference can lead to
subtle bugs because conceptually functions are often treated as a black
box and assume that they can affect only the enclosing scope by
returning a variable. With pass by reference, the argument object is
modified, even if it may not be returned by the function.

Signed-off-by: Jinzhi Chen <nxtjinzhi@gmail.com>