add Immediately Invoked Function Expressions specification
the beauty of the IIFE is that it uses a unary expression to coerce a
function declaration, which would normally need
to be explicitly called into a function expression that can
self-execute. Internally, JavaScript is running a unary
operation on the function declaration. The result of that operation
is the function expression, which is immediately
invoked with the trailing parentheses (). Besides being elegant code,
the IIFE also affords the following:
 1> It provides a closure that prevents naming conflicts.
 2> It provides elegant block scoping.
 3> It prevents pollution of the global namespace.
 4> It promotes the developer to think in terms of modular code.

Signed-off-by: Jinzhi Chen <nxtjinzhi@gmail.com>